backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-resized==1.0.2
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.0
geographiclib==2.0
geopy==2.4.1
idna==3.7
Markdown==3.6
pillow==10.4.0
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
